+++ hidive.js
@@ -43,7 +43,7 @@ const shlp    = require('sei-helper');
 const yargs   = require('yargs');
 const request = require('request');
 const agent   = require('socks5-https-client/lib/Agent');
-const vtt     = require(modulesFolder,'/module.vttconvert');
+const vtt     = require(modulesFolder+'/module.vttconvert');
 
 // m3u8
 const m3u8     = require('m3u8-parsed');
@@ -145,8 +145,11 @@ let argv = yargs
     .describe('mks','Add subtitles to mkv (if available)')
     .boolean('mks')
     .default('mks',cfg.cli.muxSubs)
+    .describe('caps','Caption subs only for non-Japanese dub / Full subs only for Japanese dub')
+    .boolean('caps')
+    .default('caps',cfg.cli.capsSubs)
     
-    .describe('a','Filenaming: Release group')
+    .describe('a','Filenaming: Release group (title tag)')
     .default('a',cfg.cli.releaseGroup)
     .describe('t','Filenaming: Series title override')
     .describe('ep','Filenaming: Episode number override (ignored in batch mode)')
@@ -154,8 +157,10 @@ let argv = yargs
     .default('suffix',cfg.cli.fileSuffix)
     
     // util
-    .describe('stag','Subtitles file: Custom title tag')
+    .describe('stag','Custom title tag in subtitle file')
     .default('stag',cfg.cli.assTitleTag)
+    .describe('ftag','Custom title tag in muxed file info')
+    .default('ftag',cfg.cli.muxTitleTag)
     .describe('nocleanup','Move temporary files to trash folder instead of deleting')
     .boolean('nocleanup')
     .default('nocleanup',cfg.cli.noCleanUp)
@@ -448,9 +453,10 @@ async function downloadMedia(videoUrl,subsUrls,fontSize){
         else{
             console.log(`[INFO] Video downloaded!\n`);
         }
-        // subs download
-        argv.stag = argv.stag ? argv.stag : argv.a
+        // stag
+        argv.stag = argv.stag ? argv.stag : argv.a;
         argv.stag = shlp.cleanupFilename(argv.stag);
+        // subs download
         let subsLangArr = Object.keys(subsUrls);
         sxList = [];
         argv.nosubs = argv.dub == 'jpn' ? false : argv.nosubs;
@@ -470,7 +476,8 @@ async function downloadMedia(videoUrl,subsUrls,fontSize){
                     sxList.push({
                         file: `${subFn}.${assExt}`,
                         language: subsLangArr[z],
-                        langCode: getLangCode(subsLangArr[z])
+                        langCode: getLangCode(subsLangArr[z]),
+                        isCaps: subsLangArr[z].match(/caps$/i) ? true : false
                     });
                     console.log(`[INFO] Subtitle downloaded and converted: ${subFn}.${assExt}`);
                 }
@@ -496,7 +503,15 @@ function getSubsUrl(file){
 async function muxStreams(){
     // fix variables
     argv.dub = argv.dub.match('-') ? argv.dub.split('-')[0] : argv.dub;
-    let addSubs = argv.mks && sxList.length > 0 && !argv.mp4 ? true : false;
+    const addSubs = argv.mks && sxList.length > 0 && !argv.mp4 ? true : false;
+    // ftag
+    argv.ftag = argv.ftag ? argv.ftag : argv.a;
+    argv.ftag = shlp.cleanupFilename(argv.ftag);
+    // caps
+    const capsOpt = !argv.caps 
+        || argv.caps &&  sxList[t].isCaps && argv.dub != 'jpn'
+        || argv.caps && !sxList[t].isCaps && argv.dub == 'jpn'
+        ? true : false;
     // check exec
     if( !argv.mp4 && !isFile(cfg.bin.mkvmerge) && !isFile(cfg.bin.mkvmerge+`.exe`) ){
         console.log(`[WARN] MKVMerge not found, skip using this...`);
@@ -509,11 +524,13 @@ async function muxStreams(){
     // mux to mkv
     if(!argv.mp4 && cfg.bin.mkvmerge){
         let mkvmux  = `-o "${fnOutput}.mkv" --disable-track-statistics-tags --engage no_variable_data `;
-            mkvmux += `--track-name "1:[${argv.a}]" --language "0:${argv.dub}" --video-tracks 1 --audio-tracks 0 --no-subtitles --no-attachments `;
+            mkvmux += `--track-name "1:[${argv.ftag}]" --language "0:${argv.dub}" --video-tracks 1 --audio-tracks 0 --no-subtitles --no-attachments `;
             mkvmux += `"${fnOutput}.ts" `;
             if(addSubs){
                 for(let t in sxList){
-                    mkvmux += `--track-name "0:${sxList[t].language}" --language "0:${sxList[t].langCode}" --default-track "0:no" "${sxList[t].file}" `;
+                    if(capsOpt){
+                        mkvmux += `--track-name "0:${sxList[t].language}" --language "0:${sxList[t].langCode}" --default-track "0:no" "${sxList[t].file}" `;
+                    }
                 }
             }
         shlp.exec(`mkvmerge`,`"${cfg.bin.mkvmerge}"`,mkvmux);
@@ -522,9 +539,11 @@ async function muxStreams(){
         let ffsubs = {fsubs:'',meta1:'',meta2:''};
         if(addSubs){
             for(let t in sxList){
-                ffsubs.fsubs += `-i "${sxList[t].file}" `
-                ffsubs.meta1 += `-map ${(parseInt(t)+1)} -c:s copy `;
-                ffsubs.meta2 += `-metadata:s:s:${(t)} title="${sxList[t].language}" -metadata:s:s:${(t)} language=${sxList[t].langCode} `;
+                if(capsOpt){
+                    ffsubs.fsubs += `-i "${sxList[t].file}" `
+                    ffsubs.meta1 += `-map ${(parseInt(t)+1)} -c:s copy `;
+                    ffsubs.meta2 += `-metadata:s:s:${(t)} title="${sxList[t].language}" -metadata:s:s:${(t)} language=${sxList[t].langCode} `;
+                }
             }
         }
         let ffext = !argv.mp4 ? `mkv` : `mp4`;
@@ -533,7 +552,7 @@ async function muxStreams(){
             ffmux += `-map 0 -c:v copy -c:a copy `;
             ffmux += ffsubs.meta1;
             ffmux += `-metadata encoding_tool="no_variable_data" `;
-            ffmux += `-metadata:s:v:0 title="[${argv.a}]" -metadata:s:a:0 language=${argv.dub} `;
+            ffmux += `-metadata:s:v:0 title="[${argv.ftag}]" -metadata:s:a:0 language=${argv.dub} `;
             ffmux += ffsubs.meta2;
             ffmux += `"${fnOutput}.${ffext}"`;
         // mux to mkv